fix: reduce Maven connector duplication
What:
- Extract shared Maven BTM parameter mapping into common Mojo base classes.
- Keep combined Maven analyze goals from deleting the Analysis Store before semantic enrichment.
- Centralize repeated Maven test setup for project, session, source, log, and reflection helpers.

Why:
- SonarCloud reported excessive new-code duplication on PR #96.
- The combined analyze flow must keep generated Analysis Store content available for Joern import.

Changes:
- Add AbstractBtmGenerationMojo and AbstractAggregateBtmGenerationMojo for shared Maven parameters and reactor roots.
- Refactor BtmGenMojo, BtmGenAggregateMojo, AnalyzeMojo, and AnalyzeAggregateMojo onto the shared mapping.
- Add shared analysis-store validation support and cleanup-policy regression coverage.
- Update Maven parity tests to handle inherited Mojo fields.

Impact:
- Fixes behavior where forensics:analyze and forensics:analyze-aggregate could honor DELETE_ON_SUCCESS before semantic enrichment.
- No public Gradle or Maven goal names were changed.
- No dependency, Java, Gradle, JaCoCo, or Sonar configuration changes were introduced.

Testing:
- .\gradlew.bat test --tests '*MavenAnalysisGoalsTest' --tests '*MavenReactorAggregationTest' --tests '*MavenBtmGenParametersTest' --tests '*BtmGenMojoTest' --tests '*MavenAnalysisStoreParityTest' --tests '*MavenJoernConfigurationParityTest' --dependency-verification strict --console=plain --stacktrace
- .\gradlew.bat test --tests '*BtmGenerationAdapterValidationTest' --dependency-verification strict --console=plain --stacktrace
- .\gradlew.bat clean test jacocoTestReport jacocoTestCoverageVerification checkPackageCoverage --dependency-verification strict --console=plain --stacktrace
- .\gradlew.bat validatePlugins --dependency-verification strict --no-daemon --console=plain --stacktrace
- .\gradlew.bat test --tests '*BtmGenerationAdapterValidationTest' --tests '*BuildToolConnectorParityTest' --tests '*MavenReactorAggregationTest' --tests '*HexagonRulesTest' --dependency-verification strict --console=plain --stacktrace
- .\gradlew.bat test --dependency-verification strict --console=plain --stacktrace
- .\gradlew.bat generateMavenPluginDescriptor --dependency-verification strict --console=plain --stacktrace
